test(sessions): satisfy lifecycle lint boundaries
